Doors FINALLY work

+++ srcs/bonus/find_wall_distance_bonus.c
@@ -18,7 +18,11 @@ int	check_if_ray_hits_door_horiz(t_data *data, double ray_angle)
 		temp_coord[0] = get_down_left_door_x(temp_coord, start_crd, ray_angle);
 	else
 		temp_coord[0] = get_up_left_door_x(temp_coord, start_crd, ray_angle);
-	if (temp_coord[0] / IMG_SIZE == start_crd[0] / IMG_SIZE)
+
+	//printf("HORIZ: temp=%lld, start=%f\n", temp_coord[0] / IMG_SIZE, start_crd[0] / IMG_SIZE);
+
+
+	if (temp_coord[0] / IMG_SIZE == (long long)start_crd[0] / IMG_SIZE)
 	{
 		data->horizon_intersection_coord[0] = temp_coord[0];
 		data->horizon_intersection_coord[1] = temp_coord[1];
@@ -47,7 +51,9 @@ int	check_if_ray_hits_door_vert(t_data *data, double ray_angle)
 	else
 		temp_coord[1] = get_up_left_door_y(temp_coord, start_crd, ray_angle);
 
-	if (temp_coord[1] / IMG_SIZE == start_crd[1] / IMG_SIZE)
+	//printf("VERT: temp=%lld, start=%f\n", temp_coord[1] / IMG_SIZE, start_crd[1] / IMG_SIZE);
+
+	if (temp_coord[1] / IMG_SIZE == (long long)start_crd[1] / IMG_SIZE)
 	{
 		data->vert_intersection_coord[0] = temp_coord[0];
 		data->vert_intersection_coord[1] = temp_coord[1];
@@ -81,6 +87,7 @@ int	check_for_wall(t_data *data, long long *check_coord, char vh_flag, double ra
 			data->invalid_horiz = 1;
 		return (1);
 	}
+
 
 	if (data->map[y][x].type == '1')
 	{
@@ -133,6 +140,9 @@ long long *w_coord, int cnt)
 		w_coord[0] = get_down_left_x(data, temp_coord, ray_angle);
 	else
 		w_coord[0] = get_up_left_x(data, temp_coord, ray_angle);
+
+//	printf("W_COORD: x=%lld\n", w_coord[0]);
+
 	data->horizon_intersection_coord[0] = temp_coord[0];
 	data->horizon_intersection_coord[1] = temp_coord[1];
 }
@@ -190,7 +200,11 @@ double	find_wall_distance(t_data *data, double ray_angle, double addition)
 	while (check_for_wall(data, horiz_coords, 'H', ray_angle) != 1)
 		get_horizon_intersection(data, ray_angle, horiz_coords, i++);
 
-
+//	if (ray_angle > 10 && ray_angle < 80)
+//	{
+//		printf("Horiz INTER COORD: x=%f, y=%f\n", data->horizon_intersection_coord[0], data->horizon_intersection_coord[1]);
+//		printf("VERT INTER COORD: x=%f, y=%f\n", data->vert_intersection_coord[0], data->vert_intersection_coord[1]);
+//	}
 //	printf("inv Hor: %d, inv vert %d\n", data->invalid_horiz, data->invalid_vert);
 
 	return (compare_distance(data, ray_angle));
